pyproject.toml
[tool.poetry]
name = "catcode.backend.coreapi"
version = "0.1.0"
description = ""
authors = ["Matias Wallenius <MatiWall@live.dk>"]
readme = "README.md"
package-mode = false

[[tool.poetry.source]]
name = "private"
url = "http://pypi.mw.local"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
httpx = "^0.26.0"
gitpython = "^3.1.41"
messaging-tools = "^0.1.0"
apscheduler = "^3.10.4"
extensions-rabbitmq = "^0.1.0"
extensions-opentelemetry = "0.1.1"
extensions-configuration = "0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
